requirements.txt
pydantic-ai
google-genai
httpx
tavily-python
pydantic
python-dotenv
pytest
pytest-asyncio
